+++ packages/components/src/happiness-engineers-tray/docs/example.jsx
@@ -1,7 +1,0 @@
-import { HappinessEngineersTray } from '../index';
-
-function HappinessEngineersTrayExample() {
-	return <HappinessEngineersTray />;
-}
-
-export default HappinessEngineersTrayExample;